Use full lenght underline including the fold markers.
Ref: https://github.com/debops/docs/issues/156#issuecomment-231672591

One thing that @drybjed changed was to go from

  ```YAML
  # Main configuration [[[
  # ------------------
  ```

to:

  ```YAML
  # Main configuration [[[
  # ----------------------
  ```

This has the advantage that in case we ever manage to get syntax highlighting for the embedded RST syntax working in Vim, it will not complain about sections being to short. I tried it but I am not sure if that is possible as the `# ` in the input for the syntax highlighting in Vim would need to be stipped out and then passed to the `rst` syntax highlighting. Ref: http://vim.wikia.com/wiki/Different_syntax_highlighting_within_regions_of_a_file#Extended_version